+++ graphistry/Plottable.py
@@ -790,7 +790,6 @@ class Plottable(Protocol):
         show_collections: Optional[bool] = None,
         collections_global_node_color: Optional[str] = None,
         collections_global_edge_color: Optional[str] = None,
-        encode: bool = True,
         validate: ValidationParam = 'autofix',
         warn: bool = True
     ) -> 'Plottable':