+++ graphistry/pygraphistry.py
@@ -2282,7 +2282,6 @@ class GraphistryClient(AuthManagerProtocol):
         show_collections: Optional[bool] = None,
         collections_global_node_color: Optional[str] = None,
         collections_global_edge_color: Optional[str] = None,
-        encode: bool = True,
         validate: ValidationParam = 'autofix',
         warn: bool = True
     ):
@@ -2291,7 +2290,6 @@ class GraphistryClient(AuthManagerProtocol):
             show_collections=show_collections,
             collections_global_node_color=collections_global_node_color,
             collections_global_edge_color=collections_global_edge_color,
-            encode=encode,
             validate=validate,
             warn=warn
         )